+++ ezlocalai/Embedding.py
@@ -1,16 +1,8 @@
 from optimum.onnxruntime import ORTModelForFeatureExtraction
 from transformers import AutoTokenizer
+from ezlocalai.Helpers import get_tokens, chunk_content
 import torch
-from typing import List
-import spacy
-import tiktoken
 import os
-
-
-def get_tokens(text: str) -> int:
-    encoding = tiktoken.get_encoding("cl100k_base")
-    num_tokens = len(encoding.encode(text))
-    return num_tokens
 
 
 class Embedding:
@@ -24,21 +16,9 @@ class Embedding:
             cache_dir=os.path.join(os.getcwd(), "models"),
         )
 
-    def chunk_content(self, text: str) -> List[str]:
-        try:
-            sp = spacy.load("en_core_web_sm")
-        except:
-            spacy.cli.download("en_core_web_sm")
-            sp = spacy.load("en_core_web_sm")
-        sp.max_length = 99999999999999999999999
-        doc = sp(text)
-        sentences = list(doc.sents)
-        content_chunks = [str(sentence).strip() for sentence in sentences]
-        return content_chunks
-
     def get_embeddings(self, input):
         tokens = get_tokens(input)
-        sentences = self.chunk_content(input)
+        sentences = chunk_content(input)
         encoded_input = self.tokenizer(
             sentences, padding=True, truncation=True, return_tensors="pt"
         ).to("cuda" if torch.cuda.is_available() else "cpu")
